+++ mysys/typelib.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2000, 2017, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2000, 2018, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License, version 2.0,
@@ -308,7 +308,7 @@ static uint parse_name(const TYPELIB *lib, const char **strpos,
     Parsed set value if (*errpos == NULL), otherwise undefined
 */
 
-my_ulonglong find_set_from_flags(const TYPELIB *lib, uint default_name,
+my_ulonglong find_set_from_flags(const TYPELIB *lib, size_t default_name,
                                  my_ulonglong cur_set, my_ulonglong default_set,
                                  const char *str, uint length, char **err_pos,
                                  uint *err_len) {
